remove big tx splitting
using nats msg splitting instead

+++ drivers/mysql/driver.go
@@ -216,7 +216,7 @@ func (d *Driver) SetupNatsServer(logger hclog.Logger) (err error) {
 	nOpts := gnatsd.Options{
 		Host:       natsAddr.IP.String(),
 		Port:       natsAddr.Port,
-		MaxPayload: g.NatsMaxPayload,
+		MaxPayload: int32(g.NatsMaxPayload),
 		//HTTPPort:   8199,
 		LogFile: "/opt/log",
 		Debug:   true,
